change int to float

+++ rpi.py
@@ -34,7 +34,7 @@ while True:
     p.start(2.5) # Initialization
     data = client_socket.recv(1024)
     pwm_string = data.decode('utf-8')
-    pwm_values = list(map(int, pwm_string.split(',')))
+    pwm_values = list(map(float, pwm_string.split(',')))
     print("recieved pwm values:", pwm_values)
     SERVER_URL = ""
